requirements.txt
Telethon=1.34.0
PyYAML=6.0.1 